+++ Sources/MultiSlider+Drag.swift
@@ -33,17 +33,9 @@ extension MultiSlider: UIGestureRecognizerDelegate {
 
         let slideViewLength = slideView.bounds.size(in: orientation)
         var targetPosition = panGesture.location(in: slideView).coordinate(in: orientation)
-        let stepSizeInView = (snapStepSize / (maximumValue - minimumValue)) * slideViewLength
-
-        // snap translation to stepSizeInView
-        if snapStepSize > 0 {
-            let translationSnapped = panGesture.translation(in: slideView).coordinate(in: orientation).rounded(stepSizeInView)
-            if 0 == Int(translationSnapped) { return }
-            panGesture.setTranslation(.zero, in: slideView)
-        }
 
         // don't cross prev/next thumb and total range
-        targetPosition = boundedDraggedThumbPosition(targetPosition: targetPosition, stepSizeInView: stepSizeInView)
+        targetPosition = boundedDraggedThumbPosition(targetPosition: targetPosition)
 
         // change corresponding value
         updateDraggedThumbValue(relativeValue: targetPosition / slideViewLength)
@@ -52,14 +44,16 @@ extension MultiSlider: UIGestureRecognizerDelegate {
             self.updateDraggedThumbPositionAndLabel()
             self.layoutIfNeeded()
         }
-
-        if isContinuous { sendActions(for: [.valueChanged, .primaryActionTriggered]) }
     }
 
     /// adjusted position that doesn't cross prev/next thumb and total range
-    private func boundedDraggedThumbPosition(targetPosition: CGFloat, stepSizeInView: CGFloat) -> CGFloat {
-        var delta = snapStepSize > 0 ? stepSizeInView : thumbViews[draggedThumbIndex].frame.size(in: orientation) / 2
-        delta = keepsDistanceBetweenThumbs ? delta : 0
+    private func boundedDraggedThumbPosition(targetPosition: CGFloat) -> CGFloat {
+        var delta: CGFloat = 0 // distance between thumbs in view coordinates
+        if distanceBetweenThumbs < 0 {
+            delta = thumbViews[draggedThumbIndex].frame.size(in: orientation) / 2
+        } else if distanceBetweenThumbs > 0 && distanceBetweenThumbs < maximumValue - minimumValue {
+            delta = (distanceBetweenThumbs / (maximumValue - minimumValue)) * slideView.bounds.size(in: orientation)
+        }
         if orientation == .horizontal { delta = -delta }
         let bottomLimit = draggedThumbIndex > 0
             ? thumbViews[draggedThumbIndex - 1].center.coordinate(in: orientation) - delta
@@ -89,6 +83,7 @@ extension MultiSlider: UIGestureRecognizerDelegate {
         if (isHapticSnap && snapStepSize > 0) || relativeValue == 0 || relativeValue == 1 {
             selectionFeedbackGenerator.generateFeedback()
         }
+        if isContinuous { sendActions(for: [.valueChanged, .primaryActionTriggered]) }
     }
 
     private func updateDraggedThumbPositionAndLabel() {