+++ Sources/MultiSlider.swift
@@ -97,6 +97,18 @@ open class MultiSlider: UIControl {
         }
     }
 
+    @IBInspectable open dynamic var valueLabelColor: UIColor? {
+        didSet {
+            valueLabels.forEach { $0.textColor = valueLabelColor }
+        }
+    }
+
+    @IBInspectable open dynamic var valueLabelFont: UIFont? {
+        didSet {
+            valueLabels.forEach { $0.font = valueLabelFont }
+        }
+    }
+
     @IBInspectable open dynamic var thumbImage: UIImage? {
         didSet {
             thumbViews.forEach { $0.image = thumbImage }
@@ -156,7 +168,17 @@ open class MultiSlider: UIControl {
         }
     }
 
-    @IBInspectable public dynamic var keepsDistanceBetweenThumbs: Bool = true
+    /// minimal distance to keep between thumbs (half a thumb by default)
+    @IBInspectable public dynamic var distanceBetweenThumbs: CGFloat = -1
+
+    @IBInspectable public dynamic var keepsDistanceBetweenThumbs: Bool {
+        get { return distanceBetweenThumbs != 0 }
+        set {
+            if keepsDistanceBetweenThumbs != newValue {
+                distanceBetweenThumbs = newValue ? -1 : 0
+            }
+        }
+    }
 
     @objc open dynamic var valueLabelFormatter: NumberFormatter = {
         let formatter = NumberFormatter()
@@ -194,7 +216,7 @@ open class MultiSlider: UIControl {
 
     // MARK: - Overrides
 
-    open override func tintColorDidChange() {
+    override open func tintColorDidChange() {
         let thumbTint = thumbViews.map { $0.tintColor } // different thumbs may have different tints
         super.tintColorDidChange()
         trackView.backgroundColor = actualTintColor
@@ -203,7 +225,7 @@ open class MultiSlider: UIControl {
         }
     }
 
-    open override var intrinsicContentSize: CGSize {
+    override open var intrinsicContentSize: CGSize {
         let thumbSize = (thumbImage ?? defaultThumbImage)?.size ?? CGSize(width: margin, height: margin)
         switch orientation {
         case .vertical:
@@ -213,20 +235,20 @@ open class MultiSlider: UIControl {
         }
     }
 
-    open override func hitTest(_ point: CGPoint, with event: UIEvent?) -> UIView? {
+    override open func hitTest(_ point: CGPoint, with event: UIEvent?) -> UIView? {
         if isHidden || alpha == 0 { return nil }
         if clipsToBounds { return super.hitTest(point, with: event) }
         return panGestureView.hitTest(panGestureView.convert(point, from: self), with: event)
     }
 
     // swiftlint:disable:next block_based_kvo
-    open override func observeValue(forKeyPath keyPath: String?, of object: Any?, change: [NSKeyValueChangeKey: Any]?, context: UnsafeMutableRawPointer?) {
+    override open func observeValue(forKeyPath keyPath: String?, of object: Any?, change: [NSKeyValueChangeKey: Any]?, context: UnsafeMutableRawPointer?) {
         if object as? NumberFormatter === valueLabelFormatter {
             updateAllValueLabels()
         }
     }
 
-    public override init(frame: CGRect) {
+    override public init(frame: CGRect) {
         super.init(frame: frame)
         setup()
     }
@@ -242,7 +264,7 @@ open class MultiSlider: UIControl {
         }
     }
 
-    open override func prepareForInterfaceBuilder() {
+    override open func prepareForInterfaceBuilder() {
         super.prepareForInterfaceBuilder()
 
         // make visual editing easier